Only add copy constraints sometimes
