+++ executor/src/witgen/block_processor.rs
@@ -34,15 +34,7 @@ impl<'a, 'b, 'c, T: FieldElement, Q: QueryCallback<T>> BlockProcessor<'a, 'b, 'c
         fixed_data: &'a FixedData<'a, T>,
         witness_cols: &'c HashSet<PolyID>,
     ) -> Self {
-        let processor = Processor::new(
-            row_offset,
-            data,
-            mutable_state,
-            fixed_data,
-            witness_cols,
-            // TODO: Copy constraints
-            Default::default(),
-        );
+        let processor = Processor::new(row_offset, data, mutable_state, fixed_data, witness_cols);
         Self {
             processor,
             identities,
